+++ src/BoundingBox.jl
@@ -461,3 +461,17 @@ function pointcrossesboundingbox(pt, bbox::BoundingBox)
     end
     error("oh dear, something has gone wrong but I don't know what")
 end
+
+"""
+    rand(bbox::BoundingBox)
+Return a random `Point` that lies inside `bbox`.
+"""
+Base.rand(bb::BoundingBox) = Point(
+    (bb.corner2.x - bb.corner1.x) * rand() + bb.corner1.x,
+    (bb.corner2.y - bb.corner1.y) * rand() + bb.corner1.y)
+
+"""
+    in(pt, bbox::BoundingBox)
+Test whether `pt` is inside `bbox`.
+"""
+Base.in(pt::Point, bbox::BoundingBox) = isinside(pt, bbox::BoundingBox)